requirements.txt
anytree==2.12.1
asgiref==3.8.1
astroid==3.3.4
boto3==1.35.31
botocore==1.35.31
certifi==2024.8.30
diff-match-patch==20230430
dill==0.3.9
Django==4.2.16
django-cors-headers==4.4.0
django-filter==21.1
django-import-export==4.1.1
django-storages==1.14.4
djangorestframework==3.15.2
factory_boy==3.3.1
Faker==30.1.0
freezegun==1.5.1
graphene==3.3
graphene-django==3.2.2
graphene-django-filter==0.6.5
graphene-file-upload==1.3.0
graphql-core==3.2.4
graphql-relay==3.2.0
isort==5.13.2
jmespath==1.0.1
mccabe==0.7.0
platformdirs==4.3.6
promise==2.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pylint==3.3.1
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
s3transfer==0.10.2
sentry-sdk==2.15.0
six==1.16.0
sqlparse==0.5.1
stringcase==1.2.0
tablib==3.5.0
text-unidecode==1.3
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
